extracted simulation methods
